Initial rfc flow commit
